refactor: update import paths for S3 storage checks and utilities
